egui: use the new editor (#1538)
This change replaces the egui markdown code with the new egui editor.
