+++ clients/egui/Cargo.toml
@@ -5,12 +5,12 @@ edition = "2021"
 
 [dependencies]
 lb = { package = "lockbook-core", path = "../../libs/core", default-features = false, features=["native-tls"] }
+lbeditor = { package = "egui_editor", path = "../../libs/editor/egui_editor" }
 eframe = "0.19.0"
 egui-winit = "0.19.0"
 egui_extras = { version = "0.19.0", features = ["image"] }
 image = { version = "0.24", default-features = false, features = ["png", "jpeg", "gif", "webp", "bmp", "ico"] }
 dark-light = "0.2.2"
-pulldown-cmark = "0.9.2"
 serde = { version = "1.0.140", features = ["derive"] }
 serde_json = "1.0.82"
